+++ .github/actions/tpv2-integration-tests/entrypoint.sh
@@ -26,13 +26,9 @@ envsubst <../../.github/actions/tpv2-integration-tests/cdn.json >./cdn.conf
 
 ./traffic_ops_golang --cfg ./cdn.conf --dbcfg ../../.github/actions/tpv2-integration-tests/database.json > out.log 2>&1 &
 
-cd "${GITHUB_WORKSPACE}/experimental/traffic-portal/dist/traffic-portal"
-envsubst <${GITHUB_WORKSPACE}/.github/actions/tpv2-integration-tests/config.json >./config.json
-node ./server/main.js -C config.json > "${GITHUB_WORKSPACE}/tp.log" 2>&1 &
-
 timeout 3m bash <<TMOUT
-	while ! curl -k "http://localhost:4200/api/4.0/ping" >/dev/null 2>&1; do
-		echo "waiting for TP dev server to proxy TO API"
+	while ! curl -k "https://localhost:6443/api/4.0/ping" >/dev/null 2>&1; do
+		echo "waiting for TO API"
 		sleep 5
 	done
 TMOUT