+++ traffic_ops/testing/api/v3/profiles_export_test.go
@@ -54,7 +54,7 @@ func TestProfilesExport(t *testing.T) {
 						t.Run(name, func(t *testing.T) {
 							resp, reqInf, err := testCase.ClientSession.ExportProfile(testCase.EndpointID())
 							for _, check := range testCase.Expectations {
-								check(t, reqInf, resp, resp.Alerts, err)
+								check(t, reqInf, resp, tc.Alerts{}, err)
 							}
 						})
 					}