+++ traffic_ops/traffic_ops_golang/cdnfederation/cdnfederations.go
@@ -499,7 +499,7 @@ func addTenancyStmt(where string) string {
 	return where
 }
 
-func getCDNFederations(inf *api.APIInfo) ([]tc.CDNFederationV5, time.Time, int, error, error) {
+func getCDNFederations(inf *api.Info) ([]tc.CDNFederationV5, time.Time, int, error, error) {
 	tenantList, err := tenant.GetUserTenantIDListTx(inf.Tx.Tx, inf.User.TenantID)
 	if err != nil {
 		return nil, time.Time{}, http.StatusInternalServerError, nil, fmt.Errorf("getting tenant list for user: %w", err)
@@ -562,7 +562,7 @@ func getCDNFederations(inf *api.APIInfo) ([]tc.CDNFederationV5, time.Time, int, 
 }
 
 // Read handles GET requests to `cdns/{{name}}/federations`.
-func Read(inf *api.APIInfo) (int, error, error) {
+func Read(inf *api.Info) (int, error, error) {
 	api.DefaultSort(inf, "cname")
 	feds, max, code, userErr, sysErr := getCDNFederations(inf)
 	if userErr != nil || sysErr != nil {
@@ -575,7 +575,7 @@ func Read(inf *api.APIInfo) (int, error, error) {
 }
 
 // ReadID handles GET requests to `cdns/{{name}}/federations/{{ID}}`.
-func ReadID(inf *api.APIInfo) (int, error, error) {
+func ReadID(inf *api.Info) (int, error, error) {
 	feds, max, code, userErr, sysErr := getCDNFederations(inf)
 	if userErr != nil || sysErr != nil {
 		return code, userErr, sysErr
@@ -611,7 +611,7 @@ func validate(fed tc.CDNFederationV5) error {
 }
 
 // Create handles POST requests to `cdns/{{name}}/federations`.
-func Create(inf *api.APIInfo) (int, error, error) {
+func Create(inf *api.Info) (int, error, error) {
 	var fed tc.CDNFederationV5
 	if err := inf.DecodeBody(&fed); err != nil {
 		return http.StatusBadRequest, fmt.Errorf("parsing request body: %w", err), nil
@@ -632,12 +632,13 @@ func Create(inf *api.APIInfo) (int, error, error) {
 		userErr, sysErr, code := api.ParseDBError(err)
 		return code, userErr, fmt.Errorf("inserting a CDN Federation: %w", sysErr)
 	}
-
+	changeLogMsg := fmt.Sprintf("CDNFEDERATION: %s, ID:%d, ACTION: Created cdnFederation", fed.CName, fed.ID)
+	api.CreateChangeLogRawTx(api.Created, changeLogMsg, inf.User, inf.Tx.Tx)
 	return inf.WriteCreatedResponse(fed, "Federation was created", "federations/"+strconv.Itoa(fed.ID))
 }
 
 // Update handles PUT requests to `cdns/{{name}}/federations/{{id}}`.
-func Update(inf *api.APIInfo) (int, error, error) {
+func Update(inf *api.Info) (int, error, error) {
 	var fed tc.CDNFederationV5
 	if err := inf.DecodeBody(&fed); err != nil {
 		return http.StatusBadRequest, fmt.Errorf("parsing request body: %w", err), nil
@@ -671,11 +672,13 @@ func Update(inf *api.APIInfo) (int, error, error) {
 		return code, userErr, sysErr
 	}
 
+	changeLogMsg := fmt.Sprintf("CDNFEDERATION: %s, ID:%d, ACTION: Updated cdnFederation", fed.CName, id)
+	api.CreateChangeLogRawTx(api.Updated, changeLogMsg, inf.User, inf.Tx.Tx)
 	return inf.WriteSuccessResponse(fed, "Federation was updated")
 }
 
 // Delete handles DELETE requests to `cdns/{{name}}/federations/{{id}}`.
-func Delete(inf *api.APIInfo) (int, error, error) {
+func Delete(inf *api.Info) (int, error, error) {
 	id := inf.IntParams["id"]
 
 	var fed tc.CDNFederationV5
@@ -684,6 +687,7 @@ func Delete(inf *api.APIInfo) (int, error, error) {
 		userErr, sysErr, code := api.ParseDBError(err)
 		return code, userErr, sysErr
 	}
-
+	changeLogMsg := fmt.Sprintf("CDNFEDERATION:%s, ID:%d, ACTION: Deleted cdnFederation", fed.CName, fed.ID)
+	api.CreateChangeLogRawTx(api.Deleted, changeLogMsg, inf.User, inf.Tx.Tx)
 	return inf.WriteSuccessResponse(fed, "Federation was deleted")
 }