+++ traffic_ops/traffic_ops_golang/cdni/capacity.go
@@ -26,7 +26,7 @@ import (
 	"github.com/apache/trafficcontrol/v8/traffic_ops/traffic_ops_golang/api"
 )
 
-func getCapacities(inf *api.APIInfo, ucdn string) (Capabilities, error) {
+func getCapacities(inf *api.Info, ucdn string) (Capabilities, error) {
 	capRows, err := inf.Tx.Tx.Query(CapabilityQuery, FciCapacityLimits, ucdn)
 	if err != nil {
 		return Capabilities{}, fmt.Errorf("querying capabilities: %w", err)