+++ traffic_ops/traffic_ops_golang/deliveryservice/servers/servers.go
@@ -767,7 +767,7 @@ func GetReadAssigned(w http.ResponseWriter, r *http.Request) {
 	api.WriteAlertsObj(w, r, http.StatusOK, alerts, servers)
 }
 
-func read(inf *api.APIInfo) ([]tc.DSServerV4, error) {
+func read(inf *api.Info) ([]tc.DSServerV4, error) {
 	queryDataString :=
 		`,
 cg.name as cachegroup,
@@ -948,22 +948,24 @@ func (dss *TODSSDeliveryService) Read(h http.Header, useIMS bool) ([]interface{}
 (ds.id in (
 	SELECT deliveryService FROM deliveryservice_server WHERE server = :server
 ) OR ds.id in (
-	SELECT id FROM deliveryservice
-	WHERE topology in (
+	SELECT d.id FROM deliveryservice d
+	JOIN cdn c ON d.cdn_id = c.id
+	WHERE d.topology in (
 		SELECT topology FROM topology_cachegroup
 		WHERE cachegroup = (
 			SELECT name FROM cachegroup
 			WHERE id = (
 				SELECT cachegroup FROM server WHERE id = :server
-			))))) 
+			)))
+	AND d.cdn_id = (SELECT cdn_id FROM server WHERE id = :server)))
 AND
-(( 
-(SELECT (t.name = 'ORG') FROM type t JOIN server s ON s.type = t.id WHERE s.id = :server) 
-OR 
-(SELECT COALESCE(ARRAY_AGG(ssc.server_capability), '{}') 
-FROM server_server_capability ssc 
-WHERE ssc."server" = :server) 
-@> 
+((
+(SELECT (t.name = 'ORG') FROM type t JOIN server s ON s.type = t.id WHERE s.id = :server)
+OR
+(SELECT COALESCE(ARRAY_AGG(ssc.server_capability), '{}')
+FROM server_server_capability ssc
+WHERE ssc."server" = :server)
+@>
 (
 SELECT COALESCE(ds.required_capabilities, '{}')
 )))