+++ traffic_ops/traffic_ops_golang/physlocation/phys_locations_test.go
@@ -96,7 +96,7 @@ func TestGetPhysLocations(t *testing.T) {
 	mock.ExpectQuery("SELECT").WillReturnRows(rows)
 	mock.ExpectCommit()
 
-	reqInfo := api.APIInfo{Tx: db.MustBegin(), Params: map[string]string{"dsId": "1"}}
+	reqInfo := api.Info{Tx: db.MustBegin(), Params: map[string]string{"dsId": "1"}}
 	obj := TOPhysLocation{
 		api.APIInfoImpl{ReqInfo: &reqInfo},
 		tc.PhysLocationNullable{},