+++ traffic_ops/traffic_ops_golang/trafficstats/stats_summary.go
@@ -49,7 +49,7 @@ func GetStatsSummary(w http.ResponseWriter, r *http.Request) {
 	return
 }
 
-func getLastSummaryDate(w http.ResponseWriter, r *http.Request, inf *api.APIInfo) {
+func getLastSummaryDate(w http.ResponseWriter, r *http.Request, inf *api.Info) {
 	queryParamsToSQLCols := map[string]dbhelpers.WhereColumnInfo{
 		"statName": dbhelpers.WhereColumnInfo{Column: "stat_name"},
 	}
@@ -86,7 +86,7 @@ func getLastSummaryDate(w http.ResponseWriter, r *http.Request, inf *api.APIInfo
 
 }
 
-func getStatsSummary(w http.ResponseWriter, r *http.Request, inf *api.APIInfo) {
+func getStatsSummary(w http.ResponseWriter, r *http.Request, inf *api.Info) {
 	queryParamsToSQLCols := map[string]dbhelpers.WhereColumnInfo{
 		"statName":            dbhelpers.WhereColumnInfo{Column: "stat_name"},
 		"cdnName":             dbhelpers.WhereColumnInfo{Column: "cdn_name"},